+++ apps/portals-example/src/app/api/dataset/[urn]/route.ts
@@ -33,3 +33,32 @@ export const GET = withAuth(
     }
   },
 );
+
+export const POST = withAuth(
+  async (
+    req: NextRequest,
+    _auth,
+    context: { params: Promise<{ urn: string }> },
+  ) => {
+    try {
+      const params = await context.params;
+      const urn = params.urn;
+
+      if (!urn) {
+        return NextResponse.json({ error: 'urn is required' }, { status: 400 });
+      }
+
+      const body = await req.json();
+      const { filters } = body;
+      const data = await datasetApi.getDatasetData(urn, filters);
+
+      return NextResponse.json(data);
+    } catch (error) {
+      apiLogger.error('Dataset data API error:', error);
+      return NextResponse.json(
+        { error: 'Failed to fetch dataset data' },
+        { status: 500 },
+      );
+    }
+  },
+);
